+++ src/node.js
@@ -28,6 +28,8 @@ import * as config from './config';
 
 export let role;
 
+let nodesBehindProxyWithKeyOnProxy;
+
 export async function getNodeRoleFromBlockchain() {
   try {
     const nodeInfo = await tendermintNdid.getNodeInfo(config.nodeId);
@@ -81,4 +83,17 @@ export async function getNodesBehindProxyFromBlockchain({ withConfig } = {}) {
       cause: error,
     });
   }
+}
+
+export async function getNodesBehindProxyWithKeyOnProxy() {
+  if (nodesBehindProxyWithKeyOnProxy == null) {
+    nodesBehindProxyWithKeyOnProxy = await getNodesBehindProxyFromBlockchain({
+      withConfig: 'KEY_ON_PROXY',
+    });
+  }
+  return nodesBehindProxyWithKeyOnProxy;
+}
+
+export function invalidateNodesBehindProxyWithKeyOnProxyCache() {
+  nodesBehindProxyWithKeyOnProxy = null;
 }